requirements-dev.txt
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0
jupyter>=1.0.0
jupyterlab>=4.0.0
nbformat>=5.0.0
nbconvert>=7.0.0
coverage>=7.0.0
pandas-profiling>=3.6.0
